streamlit/requirements.txt
streamlit
pillow
requests
streamlit-option-menu
google-cloud-storage
PyPDF2
google-cloud-aiplatform
vertexai
